applying suggestions
