+++ instrumentation/github.com/aws/aws-sdk-go-v2/otelaws/config.go
@@ -20,7 +20,7 @@ import (
 
 type config struct {
 	TracerProvider  trace.TracerProvider
-	AttributeSetter attributesetter
+	AttributeSetter AttributeSetter
 }
 
 // Option applies an option value.
@@ -47,8 +47,8 @@ func WithTracerProvider(provider trace.TracerProvider) Option {
 }
 
 // WithAttributeSetter specifies an attribute setter function for setting service specific attributes.
-// If none is specified, the service will be determined by the Defaultattributesetter function and the corresponding attributes will be included.
-func WithAttributeSetter(attributesetter attributesetter) Option {
+// If none is specified, the service will be determined by the DefaultAttributeSetter function and the corresponding attributes will be included.
+func WithAttributeSetter(attributesetter AttributeSetter) Option {
 
 	return optionFunc(func(cfg *config) {
 		if attributesetter != nil {
