+++ instrumentation/github.com/aws/aws-sdk-go-v2/otelaws/go.mod
@@ -6,8 +6,9 @@ replace go.opentelemetry.io/contrib => ../../../../../
 
 require (
 	github.com/aws/aws-sdk-go-v2 v1.12.0
-	github.com/aws/aws-sdk-go-v2/service/dynamodb v1.11.0
+	github.com/aws/aws-sdk-go-v2/service/dynamodb v1.12.0
 	github.com/aws/smithy-go v1.9.1
+	github.com/stretchr/testify v1.7.0
 	go.opentelemetry.io/otel v1.3.0
 	go.opentelemetry.io/otel/trace v1.3.0
 )
