+++ instrumentation/github.com/aws/aws-sdk-go-v2/otelaws/test/go.mod
@@ -4,9 +4,8 @@ go 1.16
 
 require (
 	github.com/aws/aws-sdk-go-v2 v1.12.0
-	github.com/aws/aws-sdk-go-v2/service/dynamodb v1.11.0
+	github.com/aws/aws-sdk-go-v2/service/dynamodb v1.12.0
 	github.com/aws/aws-sdk-go-v2/service/route53 v1.16.0
-	github.com/aws/smithy-go v1.9.1
 	github.com/stretchr/testify v1.7.0
 	go.opentelemetry.io/contrib/instrumentation/github.com/aws/aws-sdk-go-v2/otelaws v0.27.0
 	go.opentelemetry.io/otel v1.3.0
